Updates for publication

+++ classes/Blacklist.php
@@ -20,12 +20,15 @@ class Blacklist implements IteratorAggregate
      * @param null|string $context_id if null then use actual context
      * @return Blacklist
      */
-    static public function getBlacklist($context_id = null) {
-        if($context_id == null){
-            $context_id = $_GET['cid'];
+    static public function getBlacklist($context_id = null) 
+    {
+        if($context_id == null) {
+            $context_id = Request::get('cid');
         }
+
         $return = new Blacklist();
         $return->context_id = $context_id;
+        
         return $return;
     }
 
@@ -34,9 +37,9 @@ class Blacklist implements IteratorAggregate
      *
      * @return int size of list
      */
-    public function list_size() {
-        $data = DBManager::get()->fetchAll("SELECT * FROM wp_blacklist WHERE context_id = ?",
-            array($this->context_id));
+    public function list_size() 
+    {
+        $data = DBManager::get()->fetchAll("SELECT * FROM wp_blacklist WHERE context_id = ?", array($this->context_id));
 
         return sizeof($data);
     }
@@ -47,9 +50,9 @@ class Blacklist implements IteratorAggregate
      * @param string $user_id
      * @return bool
      */
-    public function isOnList($user_id) {
-        $data = DBManager::get()->fetchAll("SELECT * FROM wp_blacklist WHERE context_id = ? AND user_id = ?",
-            array($this->context_id, $user_id));
+    public function isOnList($user_id) 
+    {
+        $data = DBManager::get()->fetchAll("SELECT * FROM wp_blacklist WHERE context_id = ? AND user_id = ?", array($this->context_id, $user_id));
 
         return sizeof($data) != 0;
     }
@@ -60,18 +63,20 @@ class Blacklist implements IteratorAggregate
      * @param string $user_id id of requested user
      * @return DateTime|null datetime of expiration or null if no no expiration set or user not on list
      */
-    public function getExpiration($user_id) {
+    public function getExpiration($user_id) 
+    {
         if (!$this->isOnList($user_id)) {
             return null;
         }
-        $data = DBManager::get()->fetchFirst("SELECT * FROM wp_blacklist WHERE context_id = ? AND user_id = ?",
-            array($this->context_id, $user_id));
+
+        $data = DBManager::get()->fetchFirst("SELECT * FROM wp_blacklist WHERE context_id = ? AND user_id = ?", array($this->context_id, $user_id));
 
         if($data['expiration'] == null) {
             return null;
         } else {
-            return new DateTime('@'.$data['expiration']);
+            return new DateTime('@' . $data['expiration']);
         }
+
     }
 
     /**
@@ -80,14 +85,15 @@ class Blacklist implements IteratorAggregate
      * @param string $user_id
      * @param int|null $expiration unix Timestamp of automatic deletion
      */
-    public function addToList($user_id, $expiration = null) {
+    public function addToList($user_id, $expiration = null) 
+    {
         if(!$this->isOnList($user_id)) {
-            DBManager::get()->execute("INSERT INTO wp_blacklist SET context_id = ?, user_id = ?, expiration = ?",
-                array($this->context_id, $user_id, $expiration));
+            DBManager::get()->execute("INSERT INTO wp_blacklist SET context_id = ?, user_id = ?, expiration = ?", array($this->context_id, $user_id, $expiration));
             $notification = new WpNotifications('insert_blacklist', $user_id, PluginEngine::getLink('WorkplaceAllocation', array(), 'show'));
             $notification->setContext($this->context_id);
             $notification->sendNotification();
         }
+
     }
 
     /**
@@ -95,14 +101,15 @@ class Blacklist implements IteratorAggregate
      *
      * @param string $user_id
      */
-    public function deleteFromList($user_id) {
+    public function deleteFromList($user_id) 
+    {
         if($this->isOnList($user_id)) {
-            DBManager::get()->execute("DELETE FROM wp_blacklist WHERE context_id = ? AND user_id = ?",
-                array($this->context_id, $user_id));
+            DBManager::get()->execute("DELETE FROM wp_blacklist WHERE context_id = ? AND user_id = ?", array($this->context_id, $user_id));
             $notification = new WpNotifications('delete_blacklist', $user_id, PluginEngine::getLink('WorkplaceAllocation', array(), 'show'));
             $notification->setContext($this->context_id);
             $notification->sendNotification();
         }
+
     }
 
 
@@ -114,19 +121,24 @@ class Blacklist implements IteratorAggregate
      *        <b>Traversable</b>
      * @since 5.0.0
      */
-    public function getIterator(){
+    public function getIterator()
+    {
         $data = DBManager::get()->fetchAll("SELECT user_id, expiration FROM wp_blacklist WHERE context_id = ?", array($this->context_id));
         $usersInList = array();
+        
         foreach($data as $uid){
             $entry = array();
-            $entry['user'] = new StudIPUser($uid['user_id']);
+            $entry['user'] = User::findFull($uid['user_id']);
+
             if($uid['expiration'] != null){
-                $entry['expiration'] = new DateTime('@'.$uid['expiration']);
+                $entry['expiration'] = new DateTime('@' . $uid['expiration']);
             } else {
                 $entry['expiration'] = null;
             }
+
             $usersInList[] = $entry;
         }
+
         return new ArrayIterator($usersInList);
     }
 }
